+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 
 - pytest integration bug fixed
 - ViewMethod copy bug fixed
+- pydantic required version increased
 
 
 1.3.0 (2021-08-13)